refactor: split 3 files that crossed the 800-line cap (#1078)
db/src/sync/audiobooks.rs (802 lines) splits into audiobooks/{mod,changed,
new,removed,shared,backfill}.rs mirroring the sync/books/ bucket split.
db/src/worker/tests.rs (919 lines) sheds its periodic_scan_tick block to a
sibling db/src/worker/periodic_scan/tests.rs, matching periodic_scan.rs's
own module shape. frontend/src/components/chip_editor.rs (1013 lines)
extracts SuggestField into suggest_field.rs and the dropdown machinery
both components share into suggestion_dropdown.rs. Pure structural
extraction — no behavior change.
