+++ frontend/src/components/chip_editor.rs
@@ -1,15 +1,13 @@
 //! Generic add/remove chip editor with a substring-match suggestion dropdown.
 //! Renders chips, then input, then optional dropdown; the consumer wraps the
 //! output in whatever flex container their layout needs. Up to
-//! [`MAX_SUGGESTIONS`] case-insensitive matches against `suggestions` show,
-//! excluding values already present. Empty `suggestions` → free-text entry.
+//! `MAX_SUGGESTIONS` (see [`super::suggestion_dropdown`]) case-insensitive
+//! matches against `suggestions` show, excluding values already present.
+//! Empty `suggestions` → free-text entry.
 
 use dioxus::prelude::*;
 
-/// Cap on suggestion rows. Five fits without scroll on a typical
-/// edit-row layout and matches the "< 5 relevant suggestions"
-/// user requirement.
-const MAX_SUGGESTIONS: usize = 5;
+use super::suggestion_dropdown::{compute_suggestions, DropdownSelectionState, SuggestionDropdown};
 
 /// Per-instance handle to the editor's own root DOM node, captured on mount
 /// so a later blur can check whether focus is still somewhere inside this
@@ -650,282 +648,6 @@ fn ChipList(
     }
 }
 
-/// Per-render selection state for the dropdown: the visible suggestion
-/// rows, whether a "+ Create" trailing row is shown, the raw text to
-/// quote inside that row, and the keyboard highlight cursor. Grouped
-/// so the dropdown signature stays compact as new pieces of selection
-/// state accrete.
-#[derive(Clone, PartialEq)]
-struct DropdownSelectionState {
-    filtered: Vec<SuggestionItem>,
-    show_create_row: bool,
-    typed: String,
-    highlight: Option<usize>,
-}
-
-/// Autocomplete dropdown — filtered suggestion rows plus an optional
-/// "+ Create" trailing row. `on_pick` fires with the chosen name.
-#[component]
-fn SuggestionDropdown(
-    selection: DropdownSelectionState,
-    dropdown_header: String,
-    testid: String,
-    on_pick: EventHandler<String>,
-) -> Element {
-    let DropdownSelectionState {
-        filtered,
-        show_create_row,
-        typed,
-        highlight,
-    } = selection;
-    let create_row_index = filtered.len();
-    rsx! {
-        ul {
-            class: "chip-editor-suggestions",
-            role: "listbox",
-            "data-testid": "{testid}",
-            if !dropdown_header.is_empty() {
-                li {
-                    class: "chip-editor-suggestions-header",
-                    aria_hidden: "true",
-                    "{dropdown_header}"
-                }
-            }
-            for (i, item) in filtered.iter().cloned().enumerate() {
-                li {
-                    key: "{i}-{item.name}",
-                    // Single-line if/else avoids nightly
-                    // clippy's `suspicious_else_formatting`
-                    // lint, which trips on the multi-line
-                    // form inside rsx attribute slots.
-                    class: if Some(i) == highlight { "chip-editor-suggestion is-active" } else { "chip-editor-suggestion" },
-                    role: "option",
-                    aria_selected: if Some(i) == highlight { "true" } else { "false" },
-                    // mousedown (not click) fires before the
-                    // input's blur, so the input keeps focus
-                    // for the next add.
-                    onmousedown: {
-                        let name = item.name.clone();
-                        move |e: Event<MouseData>| {
-                            e.prevent_default();
-                            on_pick.call(name.clone());
-                        }
-                    },
-                    span { class: "chip-editor-suggestion-name", "{item.name}" }
-                    span { class: "chip-editor-suggestion-count",
-                        if item.count == 1 { "1 book" } else { "{item.count} books" }
-                    }
-                    span { class: "chip-editor-suggestion-enter", aria_hidden: "true", "\u{21a9}" }
-                }
-            }
-            if show_create_row {
-                li {
-                    key: "__create__-{typed}",
-                    class: if Some(create_row_index) == highlight { "chip-editor-suggestion chip-editor-suggestion--create is-active" } else { "chip-editor-suggestion chip-editor-suggestion--create" },
-                    role: "option",
-                    aria_selected: if Some(create_row_index) == highlight { "true" } else { "false" },
-                    onmousedown: {
-                        let value = typed.clone();
-                        move |e: Event<MouseData>| {
-                            e.prevent_default();
-                            on_pick.call(value.clone());
-                        }
-                    },
-                    span { class: "chip-editor-suggestion-name",
-                        "+ Create "
-                        span { class: "chip-editor-suggestion-quote", "\"{typed}\"" }
-                    }
-                    span { class: "chip-editor-suggestion-enter", aria_hidden: "true", "\u{21a9}" }
-                }
-            }
-        }
-    }
-}
-
-/// Compute the ≤`MAX_SUGGESTIONS` autocomplete candidates for the current
-/// query and focus state. Returns an empty `Vec` when the pool is empty,
-/// when the query doesn't match anything, or when the input is unfocused
-/// and the query is empty (open-on-focus is suppressed by `suppress_open`).
-///
-/// Filtering uses Unicode-aware `to_lowercase()` so ASCII case variants and
-/// non-ASCII look-alikes don't slip past the dedup check.
-fn compute_suggestions(
-    suggestions: &[SuggestionItem],
-    current_values: &[String],
-    query_lc: &str,
-    focused: bool,
-    suppress_open: bool,
-) -> Vec<SuggestionItem> {
-    if suggestions.is_empty() {
-        return Vec::new();
-    }
-    let current: std::collections::HashSet<String> =
-        current_values.iter().map(|s| s.to_lowercase()).collect();
-    if query_lc.is_empty() {
-        if focused && !suppress_open {
-            suggestions
-                .iter()
-                .filter(|item| !current.contains(&item.name.to_lowercase()))
-                .take(MAX_SUGGESTIONS)
-                .cloned()
-                .collect()
-        } else {
-            Vec::new()
-        }
-    } else {
-        suggestions
-            .iter()
-            .filter(|item| {
-                let lc = item.name.to_lowercase();
-                lc.contains(query_lc) && !current.contains(&lc)
-            })
-            .take(MAX_SUGGESTIONS)
-            .cloned()
-            .collect()
-    }
-}
-
-/// Props for the [`SuggestField`] component.
-#[derive(Props, Clone, PartialEq)]
-pub struct SuggestFieldProps {
-    /// The field's bound value — both the input's live text and the
-    /// committed value. Unlike [`ChipEditorProps::values`] this is a single
-    /// value, not a list: there is no separate "commit" step, typing
-    /// directly edits the value and free text is always accepted.
-    pub value: Signal<String>,
-    /// Candidate pool; same shape as [`ChipEditorProps::suggestions`].
-    pub suggestions: ReadSignal<Vec<SuggestionItem>>,
-    /// The `<input>`'s `id`, so a sibling `<label for=...>` associates with it.
-    pub id: String,
-    /// CSS class for the `<input>`.
-    #[props(default)]
-    pub class: String,
-    /// Placeholder shown when the value is empty.
-    #[props(default)]
-    pub placeholder: String,
-    /// Per-instance testid prefix, same convention as
-    /// [`ChipEditorOptions::testid_prefix`]: the input gets
-    /// `<prefix>-input`, the dropdown `<prefix>-suggestions`.
-    pub testid_prefix: String,
-}
-
-/// Single-value counterpart to [`ChipEditor`] for fields that hold one value
-/// rather than a list (e.g. Series): the same substring-match suggestion
-/// dropdown, but picking a row overwrites the field instead of adding a
-/// chip, and there is no "+ Create" row since typing already edits the
-/// value directly (free text is accepted as-is).
-#[component]
-pub fn SuggestField(props: SuggestFieldProps) -> Element {
-    let mut value = props.value;
-    let mut highlight = use_signal::<Option<usize>>(|| None);
-    // Explicit open-state, mirroring how `ChipEditor` uses `focused` +
-    // `suppress_open` to gate its dropdown. `filtered` being non-empty is
-    // *not* sufficient here: after a pick the value is set to the chosen
-    // name, which is often a substring of other pool entries, so the filter
-    // would keep matching and the dropdown would never close. `open` is set
-    // false on pick / Enter / Escape and back true on focus / input.
-    let mut open = use_signal(|| false);
-
-    let filtered = {
-        let suggestions = props.suggestions.read();
-        let query_lc = value().trim().to_lowercase();
-        compute_suggestions(&suggestions, &[value()], &query_lc, open(), false)
-    };
-    let total = filtered.len();
-    let testid_prefix = props.testid_prefix.clone();
-
-    rsx! {
-        div { class: "chip-editor-input-wrap",
-            input {
-                id: props.id.clone(),
-                class: "{props.class}",
-                "data-testid": "{testid_prefix}-input",
-                placeholder: "{props.placeholder}",
-                value: "{value}",
-                onfocus: move |_| open.set(true),
-                onblur: move |_| {
-                    open.set(false);
-                    highlight.set(None);
-                },
-                oninput: move |e| {
-                    value.set(e.value());
-                    highlight.set(None);
-                    open.set(true);
-                },
-                onkeydown: move |e: Event<KeyboardData>| {
-                    dispatch_suggest_field_keydown(e, &filtered, &mut value, &mut highlight, &mut open, total);
-                },
-            }
-            if open() && !filtered.is_empty() {
-                SuggestionDropdown {
-                    selection: DropdownSelectionState {
-                        filtered: filtered.clone(),
-                        show_create_row: false,
-                        typed: value(),
-                        highlight: highlight(),
-                    },
-                    dropdown_header: String::new(),
-                    testid: format!("{testid_prefix}-suggestions"),
-                    on_pick: move |name: String| {
-                        value.set(name);
-                        highlight.set(None);
-                        // Close on pick: without this the just-filled value
-                        // (a substring of other pool entries) keeps the
-                        // filter non-empty and the dropdown open.
-                        open.set(false);
-                    },
-                }
-            }
-        }
-    }
-}
-
-/// Arrow/Enter/Escape handling for [`SuggestField`] — a smaller sibling of
-/// [`dispatch_keydown`] without the chip-list "+ Create" row: ↑/↓ move the
-/// highlight, Enter overwrites `value` with the highlighted row and closes
-/// the dropdown (or, with nothing highlighted, leaves the typed text — it's
-/// already the live value), Escape drops the highlight and closes.
-fn dispatch_suggest_field_keydown(
-    e: Event<KeyboardData>,
-    filtered: &[SuggestionItem],
-    value: &mut Signal<String>,
-    highlight: &mut Signal<Option<usize>>,
-    open: &mut Signal<bool>,
-    total: usize,
-) {
-    match e.key() {
-        Key::ArrowDown if total > 0 => {
-            e.prevent_default();
-            let next = match highlight() {
-                Some(i) if i + 1 < total => Some(i + 1),
-                _ => Some(0),
-            };
-            highlight.set(next);
-        }
-        Key::ArrowUp if total > 0 => {
-            e.prevent_default();
-            let next = match highlight() {
-                Some(0) | None => Some(total - 1),
-                Some(i) => Some(i - 1),
-            };
-            highlight.set(next);
-        }
-        Key::Enter => {
-            if let Some(item) = highlight().and_then(|idx| filtered.get(idx)) {
-                e.prevent_default();
-                value.set(item.name.clone());
-                highlight.set(None);
-                open.set(false);
-            }
-        }
-        Key::Escape => {
-            highlight.set(None);
-            open.set(false);
-        }
-        _ => {}
-    }
-}
-
 /// Helper: dedup a list of `(name, count)` pairs into a sorted
 /// suggestion pool. Case-insensitive on the name (first-seen casing
 /// wins); when duplicates collide, the higher count wins so a
@@ -978,36 +700,5 @@ mod tests {
         ]);
         let names: Vec<&str> = out.iter().map(|i| i.name.as_str()).collect();
         assert_eq!(names, vec!["Ada", "Mira", "Zelda"]);
-    }
-
-    #[test]
-    fn compute_suggestions_returns_empty_when_pool_is_empty() {
-        let result = compute_suggestions(&[], &[], "", true, false);
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn compute_suggestions_filters_already_chosen_values() {
-        let pool = vec![
-            SuggestionItem::new("Ada", 1),
-            SuggestionItem::new("Mira", 2),
-        ];
-        let result = compute_suggestions(&pool, &["Ada".to_string()], "", true, false);
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0].name, "Mira");
-    }
-
-    #[test]
-    fn compute_suggestions_returns_empty_when_unfocused_and_query_empty() {
-        let pool = vec![SuggestionItem::new("Ada", 1)];
-        let result = compute_suggestions(&pool, &[], "", false, false);
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn compute_suggestions_returns_empty_when_suppress_open_is_true() {
-        let pool = vec![SuggestionItem::new("Ada", 1)];
-        let result = compute_suggestions(&pool, &[], "", true, true);
-        assert!(result.is_empty());
     }
 }
