+++ frontend/src/components/mod.rs
@@ -44,6 +44,16 @@ pub use format_switcher::SendToKoboButton;
 pub mod chip_editor;
 pub use chip_editor::ChipEditor;
 
+// Autocomplete-dropdown machinery (selection state, the dropdown component,
+// the substring filter) shared by `chip_editor` and `suggest_field` — split
+// out of `chip_editor.rs` once the file crossed the file-shape cap.
+mod suggestion_dropdown;
+
+// Single-value sibling of `chip_editor`'s dropdown (e.g. the Series field) —
+// split out of `chip_editor.rs` once the file crossed the file-shape cap.
+pub mod suggest_field;
+pub use suggest_field::SuggestField;
+
 pub mod atrium;
 
 // Shared loading/error/not-found states for top-level pages — see
